Code compiles again. Still not functioning. Why code complete?

+++ include/tcp-disrupt.h
@@ -25,6 +25,6 @@
 /* Displays the usage string for this program */
 void display_usage(char *name);
 
-void disrupt_session(char *sourceIP, uint16_t sourcePort, char *destinationIP, uint16_t destinationPort, uint32_t sequenceNumber, uint16_t ackNumber);
+void disrupt_session(char *sourceIP, uint16_t sourcePort, char *destinationIP, uint16_t destinationPort, uint32_t sequenceNumber, uint32_t ackNumber);
 
 #endif
